# cc-registry-v2/backend/openapi.yaml
openapi: 3.1.0
info:
  title: RunWhen CodeCollection Registry API
  description: |
    Backend API for the RunWhen CodeCollection Registry. Manages codecollections,
    codebundles, tasks, documentation, AI enhancement, Helm charts, and chat.

    ## Architecture

    This API is the single source of truth for all registry data. It is consumed by:
    - **Frontend** — React UI at the same origin
    - **MCP Server** — stateless tool server for LLM clients (via `REGISTRY_API_URL`)
    - **Celery Workers** — background task processing (sync, parse, enhance, index)

    ## Authentication

    Admin endpoints require a Bearer token in the `Authorization` header.
    Public endpoints (registry, chat, analytics) require no authentication.

    ## Data Flow

    ```
    codecollections.yaml → Seed → Sync (git clone) → Parse (codebundles) → Enhance (AI) → Serve
    ```

  version: 2.0.0
  contact:
    name: RunWhen
    url: https://runwhen.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8001
    description: Local development
  - url: http://cc-registry-backend:8001
    description: In-cluster (Kubernetes)
...
paths:
...
  /api/v1/admin/collections:
    get:
      tags: [Admin - Collections]
      summary: List collections (admin)
      operationId: adminListCollections
      security: [{ BearerAuth: [] }]
      parameters:
        - name: skip
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Collection list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/CollectionWithStats" } }
    post:
      tags: [Admin - Collections]
      summary: Create collection
      operationId: createCollection
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionCreate"
      responses:
        "200":
          description: Created collection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionWithStats" }
  /api/v1/admin/collections/{collection_id}:
    get:
      tags: [Admin - Collections]
      summary: Get collection by ID (admin)
      operationId: adminGetCollection
      security: [{ BearerAuth: [] }]
      parameters:
        - name: collection_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Collection details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionWithStats" }
    put:
      tags: [Admin - Collections]
      summary: Update collection
      operationId: updateCollection
      security: [{ BearerAuth: [] }]
      parameters:
        - name: collection_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionWithStats" }
    delete:
      tags: [Admin - Collections]
      summary: Delete collection
      operationId: deleteCollection
      security: [{ BearerAuth: [] }]
      parameters:
        - name: collection_id
          in: path
          required: true
          schema: { type: integer }
        - name: hard_delete
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  collection_slug: { type: string }
  /api/v1/admin/collections/{collection_id}/sync:
    post:
      tags: [Admin - Collections]
      summary: Sync collection from git
      operationId: syncCollection
      security: [{ BearerAuth: [] }]
      parameters:
        - name: collection_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sync task submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskSubmitted" }
  /api/v1/admin/collections/{collection_id}/parse:
    post:
      tags: [Admin - Collections]
      summary: Parse collection codebundles
      operationId: parseCollection
      security: [{ BearerAuth: [] }]
      parameters:
        - name: collection_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Parse task submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskSubmitted" }
...
  /api/v1/admin/seed-from-yaml:
    post:
      tags: [Admin - Operations]
      summary: Seed database from YAML
      operationId: seedFromYaml
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Seed task submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskSubmitted" }
  /api/v1/admin/validate-yaml-seed:
    post:
      tags: [Admin - Operations]
      summary: Validate YAML entries
      operationId: validateYamlSeed
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Validation task submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskSubmitted" }

  # =========================================================================
  # Admin - Inventory
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin token (must start with `admin-`)
...
  schemas:
...
    TaskSubmitted:
      type: object
      properties:
        message: { type: string }
        task_id: { type: string, format: uuid }
        status: { type: string, example: submitted }
    CollectionCreate:
      type: object
      required: [name, slug, git_url]
      properties:
        name: { type: string }
        slug: { type: string }
        git_url: { type: string, format: uri }
        description: { type: string }
        owner: { type: string }
        owner_email: { type: string, format: email }
        owner_icon: { type: string, format: uri }
        git_ref: { type: string, default: main }
        is_active: { type: boolean, default: true }
    CollectionUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        git_url: { type: string, format: uri }
        git_ref: { type: string }
        owner: { type: string }
        owner_email: { type: string }
        owner_icon: { type: string }
        is_active: { type: boolean }
    CollectionWithStats:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, example: AWS Cloud/Custodian CodeCollection }
        slug: { type: string, example: aws-c7n-codecollection }
        description: { type: string }
        git_url: { type: string, format: uri }
        git_ref: { type: string }
        owner: { type: string }
        owner_email: { type: string }
        owner_icon: { type: string }
        is_active: { type: boolean }
        codebundle_count: { type: integer }
        task_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...